requirements.txt
Flask
pymysql
cryptography
pandas
plotly
python-dotenv
gunicorn
numpy